+++ backend/flask-api-serverless/energy.py
@@ -23,7 +23,7 @@ def getEnergy(data, energy_level, num_results):
 
     
     df = pd.DataFrame(data)
-    df['distances'] = abs(df["energy"] - energy_level)
+    df['distances'] = abs(df["energy"].astype(float) - float(energy_level))
     df.sort_values("distances", ascending=True,  inplace=True)
     idx = df.columns.get_loc("distances")
     df.drop(df.columns[idx], axis=1, inplace=True)
