+++ backend/flask-api-serverless/liveness.py
@@ -23,7 +23,7 @@ def getLiveness(data, live_level, num_results):
 
     
     df = pd.DataFrame(data)
-    df['distances'] = abs(df["liveness"] - live_level)
+    df['distances'] = abs(df["liveness"].astype(float) - float(live_level))
     df.sort_values("distances", ascending=True,  inplace=True)
     idx = df.columns.get_loc("distances")
     df.drop(df.columns[idx], axis=1, inplace=True)
