+++ backend/flask-api-serverless/popularity.py
@@ -29,7 +29,7 @@ def getPopularity(data, popularity_level, num_results):
 
     
     df = pd.DataFrame(data)
-    df['distances'] = abs(df["popularity"] - popularity_level)
+    df['distances'] = abs(df["popularity"].astype(float) - float(popularity_level))
     df.sort_values("distances", ascending=True,  inplace=True)
     idx = df.columns.get_loc("distances")
     df.drop(df.columns[idx], axis=1, inplace=True)
